Custom upgrade bool for Edgeware accounts migration
