+++ modules/edge-voting/Cargo.toml
@@ -8,15 +8,15 @@ edition = "2018"
 serde = { version = "1.0", default-features = false, optional = true }
 safe-mix = { version = "1.0", default-features = false }
 codec = { package = "parity-scale-codec", version = "1.0.6", default-features = false, features = ["derive"] }
-sp-std = { git = "https://github.com/paritytech/substrate.git", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", default-features = false }
-frame-support = { git = "https://github.com/paritytech/substrate.git", default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate.git", default-features = false }
-pallet-balances = { git = "https://github.com/paritytech/substrate.git", default-features = false }
+sp-std = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
+sp-runtime = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
+frame-support = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
+frame-system = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
+pallet-balances = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
 
 [dev-dependencies]
-sp-io = { git = "https://github.com/paritytech/substrate.git", default-features = false }
-sp-core = { git = "https://github.com/paritytech/substrate.git", default-features = false }
+sp-io = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
+sp-core = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup", default-features = false }
 
 [features]
 default = ["std"]
