+++ node/executor/Cargo.toml
@@ -8,29 +8,29 @@ edition = "2018"
 [dependencies]
 trie-root = "0.16.0"
 codec = { package = "parity-scale-codec", version = "1.0.0" }
-sp-io = { git = "https://github.com/paritytech/substrate.git" }
-sp-state-machine = { git = "https://github.com/paritytech/substrate.git" }
-sc-executor = { git = "https://github.com/paritytech/substrate.git" }
-sp-core = { git = "https://github.com/paritytech/substrate.git" }
-sp-trie = { git = "https://github.com/paritytech/substrate.git" }
-frame-benchmarking = { git = "https://github.com/paritytech/substrate.git" }
+sp-io = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-state-machine = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-executor = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-core = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-trie = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+frame-benchmarking = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
 
 edgeware-primitives = { path = "../primitives" }
 edgeware-runtime = { path = "../runtime" }
 
 [dev-dependencies]
-test-client = { package = "substrate-test-client", git = "https://github.com/paritytech/substrate.git" }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git" }
-frame-support = { git = "https://github.com/paritytech/substrate.git" }
-pallet-balances = { git = "https://github.com/paritytech/substrate.git" }
-pallet-transaction-payment = { git = "https://github.com/paritytech/substrate.git" }
-pallet-session = { git = "https://github.com/paritytech/substrate.git" }
-frame-system = { git = "https://github.com/paritytech/substrate.git" }
-pallet-timestamp = { git = "https://github.com/paritytech/substrate.git" }
-pallet-treasury = { git = "https://github.com/paritytech/substrate.git" }
-pallet-contracts = { git = "https://github.com/paritytech/substrate.git" }
-pallet-grandpa = { git = "https://github.com/paritytech/substrate.git" }
-pallet-indices = { git = "https://github.com/paritytech/substrate.git" }
+test-client = { package = "substrate-test-client", git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-runtime = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+frame-support = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-balances = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-transaction-payment = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-session = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+frame-system = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-timestamp = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-treasury = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-contracts = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-grandpa = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-indices = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
 wabt = "0.9.2"
 criterion = "0.3.0"
