+++ node/inspect/Cargo.toml
@@ -8,10 +8,10 @@ edition = "2018"
 codec = { package = "parity-scale-codec", version = "1.0.0" }
 derive_more = "0.99"
 log = "0.4.8"
-sc-cli = { git = "https://github.com/paritytech/substrate.git" }
-sc-client-api = { git = "https://github.com/paritytech/substrate.git" }
-sc-service = { default-features = false, git = "https://github.com/paritytech/substrate.git" }
-sp-blockchain = { git = "https://github.com/paritytech/substrate.git" }
-sp-core = { git = "https://github.com/paritytech/substrate.git" }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git" }
+sc-cli = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-client-api = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-service = { default-features = false, git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-blockchain = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-core = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-runtime = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
 structopt = "0.3.8"
