+++ node/rpc/Cargo.toml
@@ -6,16 +6,16 @@ edition = "2018"
 
 [dependencies]
 jsonrpc-core = "14.0.3"
-sp-api = { git = "https://github.com/paritytech/substrate.git" }
-sc-client = { git = "https://github.com/paritytech/substrate.git" }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git" }
-pallet-contracts-rpc = { git = "https://github.com/paritytech/substrate.git" }
-pallet-transaction-payment-rpc = { git = "https://github.com/paritytech/substrate.git" }
-substrate-frame-rpc-system = { git = "https://github.com/paritytech/substrate.git" }
-sp-transaction-pool = { git = "https://github.com/paritytech/substrate.git" }
-sc-keystore = { git = "https://github.com/paritytech/substrate.git" }
-sp-consensus = { git = "https://github.com/paritytech/substrate.git" }
-sp-blockchain = { git = "https://github.com/paritytech/substrate.git" }
+sp-api = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-client = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-runtime = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-contracts-rpc = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-transaction-payment-rpc = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+substrate-frame-rpc-system = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-transaction-pool = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-keystore = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-consensus = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-blockchain = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
 
 edgeware-primitives = { path = "../primitives" }
 edgeware-runtime = { path = "../runtime" }
