+++ node/runtime/src/lib.rs
@@ -128,7 +128,7 @@ impl frame_system::Trait for Runtime {
 	type ModuleToIndex = ModuleToIndex;
 	type AccountData = pallet_balances::AccountData<Balance>;
 	type OnNewAccount = ();
-	type OnReapAccount = (Balances, Staking, Contracts, Session);
+	type OnKilledAccount = ();
 }
 
 parameter_types! {
