+++ node/testing/Cargo.toml
@@ -6,25 +6,25 @@ description = "Test utilities for Edgeware."
 edition = "2018"
 
 [dependencies]
-pallet-balances = { git = "https://github.com/paritytech/substrate.git" }
-sc-client = { git = "https://github.com/paritytech/substrate.git" }
+pallet-balances = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-client = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
 codec = { package = "parity-scale-codec", version = "1.0.0" }
-pallet-contracts = { git = "https://github.com/paritytech/substrate.git" }
-pallet-grandpa = { git = "https://github.com/paritytech/substrate.git" }
-pallet-indices = { git = "https://github.com/paritytech/substrate.git" }
-sp-keyring = { git = "https://github.com/paritytech/substrate.git" }
-sp-core = { git = "https://github.com/paritytech/substrate.git" }
-sp-io = { git = "https://github.com/paritytech/substrate.git" }
-frame-support = { git = "https://github.com/paritytech/substrate.git" }
-pallet-session = { git = "https://github.com/paritytech/substrate.git" }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git" }
-pallet-staking = { git = "https://github.com/paritytech/substrate.git" }
-sc-executor = { git = "https://github.com/paritytech/substrate.git" }
-frame-system = { git = "https://github.com/paritytech/substrate.git" }
-substrate-test-client = { git = "https://github.com/paritytech/substrate.git" }
-pallet-timestamp = { git = "https://github.com/paritytech/substrate.git" }
-pallet-transaction-payment = { git = "https://github.com/paritytech/substrate.git" }
-pallet-treasury = { git = "https://github.com/paritytech/substrate.git" }
+pallet-contracts = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-grandpa = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-indices = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-keyring = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-core = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-io = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+frame-support = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-session = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sp-runtime = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-staking = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+sc-executor = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+frame-system = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+substrate-test-client = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-timestamp = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-transaction-payment = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
+pallet-treasury = { git = "https://github.com/drewstone/substrate.git", branch = "drew.customup" }
 wabt = "0.9.2"
 
 edgeware-executor = { path = "../executor" }
