+++ node/transaction-factory/src/lib.rs
@@ -83,7 +83,7 @@ pub fn factory<RA, Backend, Exec, Block, RtApi, Sc>(
 	mut factory_state: RA,
 	client: &Arc<Client<Backend, Exec, Block, RtApi>>,
 	select_chain: &Sc,
-) -> sc_cli::error::Result<()>
+) -> sc_cli::Result<()>
 where
 	Block: BlockT,
 	Exec: sc_client::CallExecutor<Block, Backend = Backend> + Send + Sync + Clone,
@@ -97,7 +97,7 @@ where
 	RA: RuntimeAdapter<Block = Block>,
 	Block::Hash: From<sp_core::H256>,
 {
-	let best_header: Result<<Block as BlockT>::Header, sc_cli::error::Error> =
+	let best_header: Result<<Block as BlockT>::Header, sc_cli::Error> =
 		select_chain.best_chain().map_err(|e| format!("{:?}", e).into());
 	let mut best_hash = best_header?.hash();
 	let mut best_block_id = BlockId::<Block>::hash(best_hash);
